Revert API subpackage move
