--- docs/opentelemetry.api.loader.rst
@@ -1,4 +1,0 @@
-opentelemetry.api.loader module
-===============================
-
-.. automodule:: opentelemetry.api.loader
