--- docs/opentelemetry.api.trace.rst
@@ -1,4 +1,0 @@
-opentelemetry.api.trace package
-===============================
-
-.. automodule:: opentelemetry.api.trace
