+++ docs/opentelemetry.loader.rst
@@ -1,0 +1,4 @@
+opentelemetry.loader module
+===========================
+
+.. automodule:: opentelemetry.loader
