+++ docs/opentelemetry.trace.rst
@@ -1,0 +1,4 @@
+opentelemetry.trace package
+===========================
+
+.. automodule:: opentelemetry.trace
