+++ opentelemetry-api/setup.py
@@ -16,8 +16,7 @@ import os
 import setuptools
 
 BASE_DIR = os.path.dirname(__file__)
-VERSION_FILENAME = os.path.join(BASE_DIR, "src", "opentelemetry", "api",
-                                "version.py")
+VERSION_FILENAME = os.path.join(BASE_DIR, "src", "opentelemetry", "version.py")
 PACKAGE_INFO = {}
 with open(VERSION_FILENAME) as f:
     exec(f.read(), PACKAGE_INFO)
@@ -48,7 +47,7 @@ setuptools.setup(
     license="Apache-2.0",
     package_dir={"": "src"},
     packages=setuptools.find_namespace_packages(where="src",
-                                                include="opentelemetry.api.*"),
+                                                include="opentelemetry.*"),
     url=("https://github.com/open-telemetry/opentelemetry-python"
          "/tree/master/opentelemetry-api"),
     zip_safe=False,
