+++ opentelemetry-api/src/opentelemetry/resources/__init__.py
