+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/__init__.py
@@ -12,9 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from opentelemetry import api
+from opentelemetry import trace as trace_api
 
 
-class Tracer(api.trace.Tracer):
+class Tracer(trace_api.Tracer):
     pass
-
